// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Gender = {
  Female: 'Female',
  Male: 'Male'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const Role = {
  Admin: 'Admin',
  User: 'User'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const TokenType = {
  Bearer: 'Bearer'
} as const

export type TokenType = (typeof TokenType)[keyof typeof TokenType]